Import node-fetch dynamically
Since v3.0.0 node-fetch is an es-module only package, and cannot be
imported with require, so using the dynamic import is the way to go in a
library like this, which is transpiled to commonjs.

The actual import is only done the first time, all the subsequent calls
will immediately resolve, so it shouldn't be a performance issue or
anything like that.